// Frontend/dr-one-portal/campaign-analytics/src/views/aggregate/models/CampaignMetrics.model.ts
export interface CampaignMetrics {
  performanceType?:      string;
  price?:                number;
  impressions?:          number;
  impressionRate?:       number;
  downloads?:            number;
  totalClientsTargeted?: number;
  totalResponseCount?:   number;
  totalAudienceSize?:    number;
  clicks?:               number;
  videoViewed?:          number;
  videoViewedRate?:      number;
  eCPM?:                 number;
  ctiIu?:                number;
  apkInstalledIu?:       number;
  ctiExt?:               number;
  appInstalledExt?:      number;
  errors?:               number;
  ctr?:                  number;
}